Updated values in form when querying MOS data
